+++ app/types/fileTypes.ts
@@ -1,0 +1,18 @@
+export interface FolderCardType {
+    folderName: string,
+    folderURL: string,
+}
+
+export interface FileCardType {
+    fileName: string,
+    fileType: string,
+    fileURL: string,
+    filePreviewURL: string,
+    tags: string[],
+    lastModified: string
+}
+
+export interface FileListType {
+    folders: FolderCardType[],
+    files: FileCardType[],
+}
